product/views/product_views.py
from rest_framework import generics
from product.models import Product
from product.serializers import ProductSerializer

class ProductListCreateView(generics.ListCreateAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer
