pyproject.toml
[project]
name = "tarefa"
version = "0.1.0"
description = ""
authors = [
    {name = "WilfredoSolorzano",email = "wsolorzano00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
